+++ src/Development/Shake/Plus/Core.hs
@@ -1,7 +1,7 @@
 {- |
    Module     : Development.Shake.Plus.Core
    Copyright  : Copyright (C) 2020 Daniel Firth
-   Maintainer : Daniel Firth <dan.firth@homotopic.tech
+   Maintainer : Daniel Firth <dan.firth@homotopic.tech>
    License    : MIT
    Stability  : experimental
 
@@ -19,7 +19,6 @@ module Development.Shake.Plus.Core (
 , ShakePlus
 , runRAction
 , runShakePlus
-, runSimpleShakePlus
 , parallel
 , forP
 , par
@@ -107,14 +106,6 @@ instance MonadThrow (RAction r) where
 
 instance MonadThrow (ShakePlus r) where
   throwM = liftIO . Control.Exception.throwIO
-
--- | Run a `ShakePlus` with just a `LogFunc` in the environment that logs to stderr.
-runSimpleShakePlus :: MonadIO m => ShakePlus LogFunc a -> m ()
-runSimpleShakePlus m = do
-  lo <- logOptionsHandle stderr True
-  (lf, dlf) <- newLogFunc (setLogMinLevel LevelInfo lo)
-  liftIO $ Development.Shake.shakeArgs Development.Shake.shakeOptions $ void $ runShakePlus lf m
-  dlf
 
 -- | Unlifted `Development.Shake.parallel`.
 parallel :: MonadUnliftAction m => [m a] -> m [a]